refactor: move form plugin to dm-core-plugins
