+++ example/package.json
@@ -6,7 +6,6 @@
     "@development-framework/dm-core": "link:../packages/dm-core",
     "@development-framework/blueprint": "link:../packages/blueprint",
     "@development-framework/default-pdf": "link:../packages/default-pdf",
-    "@development-framework/form": "link:../packages/form",
     "@development-framework/header": "link:../packages/header",
     "@development-framework/dm-core-plugins": "link:../packages/dm-core-plugins",
     "@craco/craco": "^7.0.0",
