+++ example/src/plugins.js
@@ -1,7 +1,6 @@
 export default [
     import('@development-framework/blueprint'),
     import('@development-framework/default-pdf'),
-    import('@development-framework/form'),
     import('@development-framework/header'),
     import('@development-framework/dm-core-plugins'),
 ]
