+++ packages/dm-core-plugins/package.json
@@ -16,18 +16,25 @@
     "react-copy-to-clipboard": "^5.1.0",
     "react-notifications": "^1.4.3",
     "styled-components": "^5.1.0",
-    "typescript": "^4.5.5"
+    "typescript": "^4.5.5",
+    "object-path": "^0.11.8",
+    "react-hook-form": "^7.31.2"
   },
   "devDependencies": {
     "@development-framework/dm-core": ">=1.0.37",
     "@testing-library/react": "^12.1.2",
     "@testing-library/react-hooks": "^7.0.2",
+    "@testing-library/user-event": "14.4.3",
     "@types/jest": "^27.4.0",
     "@types/js-yaml": "^4.0.5",
     "@types/react": "^17.0.39",
     "@types/react-dom": "^17.0.11",
     "@types/styled-components": "^5.1.25",
-    "ts-jest": "^27.1.3"
+    "ts-jest": "^27.1.3",
+    "jest": "^27.1.3",
+    "react-router-dom": ">=5.1.2",
+    "react-test-renderer": "^17.0.2",
+    "ts-loader": "^9.3.1"
   },
   "peerDependencies": {
     "@development-framework/dm-core": ">=1.0.37",
