+++ packages/dm-core-plugins/src/form/FormComonent.tsx
@@ -2,17 +2,15 @@ import * as React from 'react'
 
 import {
   BlueprintPicker,
-  EPluginType,
   IUIPlugin,
   Loading,
-  TPlugin,
   useBlueprint,
   useDocument,
 } from '@development-framework/dm-core'
-import { Form } from './Form'
+import {Form} from './Form'
 import styled from 'styled-components'
 import TextWidget from './widgets/TextWidget'
-import { TWidget } from './types'
+import {TWidget} from './types'
 
 // The custom widgets goes under here,
 // this may at some point be moved out from the form package.
@@ -22,10 +20,10 @@ const ErrorHelperText = styled.div`
 
 const widgets = {
   TypeWidget: (props: TWidget) => {
-    const { id, namePath, label, value } = props
-    const { blueprint, isLoading } = useBlueprint(value)
+    const {id, namePath, label, value} = props
+    const {blueprint, isLoading} = useBlueprint(value)
 
-    if (isLoading) return <Loading />
+    if (isLoading) return <Loading/>
     if (blueprint === undefined) return <div>Could not find the blueprint</div>
 
     const datasourceId = value.split('/')[0]
@@ -39,7 +37,8 @@ const widgets = {
           readOnly={true}
           value={value}
           // eslint-disable-next-line @typescript-eslint/no-empty-function
-          onChange={() => {}}
+          onChange={() => {
+          }}
           onClick={() => {
             // @ts-ignore
             window
@@ -51,7 +50,7 @@ const widgets = {
     )
   },
   BlueprintPickerWidget: (props: TWidget) => {
-    const { label, variant, onChange, value, helperText } = props
+    const {label, variant, onChange, value, helperText} = props
     return (
       <>
         <BlueprintPicker
@@ -70,11 +69,11 @@ const widgets = {
   },
 }
 
-const PluginComponent = (props: IUIPlugin) => {
-  const { config, onOpen } = props
-  const { idReference } = props
+export const FormComponent = (props: IUIPlugin) => {
+  const {config, onOpen} = props
+  const {idReference} = props
   const [document, loading, updateDocument] = useDocument<any>(idReference, 999)
-  if (loading) return <Loading />
+  if (loading) return <Loading/>
 
   const handleOnSubmit = (formData: any) => {
     updateDocument(formData, true)
@@ -95,10 +94,3 @@ const PluginComponent = (props: IUIPlugin) => {
     />
   )
 }
-
-export const plugins: TPlugin[] = [
-  {
-    pluginName: 'form',
-    component: PluginComponent,
-  },
-]
