--- packages/form/jest.config.js
@@ -1,8 +1,0 @@
-const packageJson = require('./package.json')
-const base = require('../../jest.config.base')
-module.exports = {
-  ...base,
-  testEnvironment: 'jsdom',
-  name: packageJson.name,
-  displayName: packageJson.name,
-}
