--- packages/form/src/widgets/TextWidget/TextWidget.tsx
@@ -1,19 +1,0 @@
-import React from 'react'
-
-import { TextField, Icon } from '@equinor/eds-core-react'
-
-import { error_filled } from '@equinor/eds-icons'
-import { TWidget } from '../../types'
-
-Icon.add({ error_filled })
-
-const TextWidget = (props: TWidget) => {
-  const { label, onChange } = props
-
-  const _onChange = ({
-    target: { value },
-  }: React.ChangeEvent<HTMLInputElement>) => onChange(value === '' ? '' : value)
-  return <TextField {...props} onChange={_onChange} label={label} />
-}
-
-export default TextWidget
