--- packages/form/tsconfig.json
@@ -1,29 +1,0 @@
-{
-  "compilerOptions": {
-    "outDir": "dist",
-    "rootDir": "src",
-    "lib": ["dom", "dom.iterable", "esnext"],
-    "allowSyntheticDefaultImports": true,
-    "forceConsistentCasingInFileNames": true,
-    "resolveJsonModule": true,
-    "isolatedModules": true,
-    "noEmit": false,
-    "downlevelIteration": true,
-    "baseUrl": ".",
-    "declaration": true,
-    "declarationMap": true,
-    "sourceMap": true,
-    "allowJs": true,
-    "strict": true,
-    "skipLibCheck": true,
-    "jsx": "react",
-    "esModuleInterop": true,
-    "module": "commonjs",
-    "moduleResolution": "node",
-    "target": "es2022",
-    "composite": false,
-    "useUnknownInCatchVariables": false
-  },
-  "include": ["src"],
-  "exclude": ["node_modules", "dist"]
-}
